ExceptionHandler with personalized message

+++ src/main/java/com/leonardo/management/repositories/FuncionariosDatabase.java
@@ -6,6 +6,6 @@ import com.leonardo.management.entities.Funcionario;
 
 public interface FuncionariosDatabase {
 	
-	Map<Long, Funcionario> getFuncionarios();
+	Map<Integer, Funcionario> getFuncionarios();
 
 }
